test-runner/pyproject.toml
[project]
name = "ai-workflow-test-runner"
version = "1.0.0"
description = "AI Workflow System - Advanced Test Runner with rich terminal output"
authors = [
    {name = "AI Workflow Team", email = "team@aiworkflow.com"}
]
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
keywords = ["testing", "workflow", "orchestration", "monitoring"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: System :: Monitoring",
]

dependencies = [
    "rich>=13.7.0",
    "httpx>=0.25.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
workflow-test = "test_runner.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "types-pyyaml>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
# UV-specific settings
prerelease = "allow"
compile = true

[tool.uv.pip]
# Pip compatibility settings
index-url = "https://pypi.org/simple"
extra-index-url = []
find-links = []

[tool.ruff]
target-version = "py38"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "DTZ", # flake8-datetimez
    "T10", # flake8-debugger
    "RET", # flake8-return
    "SIM", # flake8-simplify
    "PTH", # flake8-use-pathlib
]
ignore = [
    "E501", # line too long (handled by black)
]

[tool.ruff.isort]
known-first-party = ["test_runner"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311", "py312"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --cov=test_runner --cov-report=term-missing"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "*/cli.py"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false